# zbputils/control/web/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "zbp restful api document",
        "title": "zbp api",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:3000",
    "basePath": "/",
    "paths": {
        "/api/deleteGroup": {
            "get": {
                "description": "删除群聊或好友",
                "tags": [
                    "通用"
                ],
                "summary": "删除群聊或好友",
                "parameters": [
                    {
                        "description": "删除群聊或好友入参",
                        "name": "object",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/types.DeleteGroupParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            }
        },
...
        "/api/getFriendList": {
            "get": {
                "description": "获取好友列表",
                "tags": [
                    "通用"
                ],
                "summary": "获取好友列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "机器人qq",
                        "name": "selfId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            }
        },
        "/api/getGroupList": {
            "get": {
                "description": "获取群列表",
                "tags": [
                    "通用"
                ],
                "summary": "获取群列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "机器人qq",
                        "name": "selfId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            }
        },
        "/api/getGroupMemberList": {
            "get": {
                "description": "获取群成员列表,groupId为0的时候拉取好友信息",
                "tags": [
                    "通用"
                ],
                "summary": "获取群成员列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "机器人qq",
                        "name": "selfId",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "群聊id",
                        "name": "groupId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            }
        },
...
        "/api/getRequestList": {
            "get": {
                "description": "获取所有的请求",
                "tags": [
                    "通用"
                ],
                "summary": "获取所有的请求",
                "parameters": [
                    {
                        "description": "获取所有的请求入参",
                        "name": "object",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/types.BotParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/types.RequestVo"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/api/handleRequest": {
            "post": {
                "description": "处理一个请求",
                "tags": [
                    "通用"
                ],
                "summary": "处理一个请求",
                "parameters": [
                    {
                        "description": "处理一个请求入参",
                        "name": "object",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/types.HandleRequestParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            }
        },
...
        "/api/login": {
            "post": {
                "description": "前端登录",
                "tags": [
                    "用户"
                ],
                "summary": "前端登录",
                "parameters": [
                    {
                        "description": "前端登录",
                        "name": "object",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/types.LoginParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/types.LoginResultVo"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/api/manage/getAllPlugin": {
            "get": {
                "description": "获取所有插件的状态",
                "tags": [
                    "插件"
                ],
                "summary": "获取所有插件的状态",
                "parameters": [
                    {
                        "description": "获取所有插件的状态入参",
                        "name": "object",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/types.AllPluginParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/types.PluginVo"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/manage/getPlugin": {
            "get": {
                "description": "获取某个插件的状态",
                "tags": [
                    "插件"
                ],
                "summary": "获取某个插件的状态",
                "parameters": [
                    {
                        "description": "获取某个插件的状态入参",
                        "name": "object",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/types.PluginParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/types.PluginVo"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/manage/updateAllPluginStatus": {
            "post": {
                "description": "更改某群所有插件状态",
                "tags": [
                    "响应"
                ],
                "summary": "更改某群所有插件状态",
                "parameters": [
                    {
                        "description": "修改插件状态入参",
                        "name": "object",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/types.AllPluginStatusParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            }
        },
        "/api/manage/updatePluginStatus": {
            "post": {
                "description": "更改某一个插件状态",
                "tags": [
                    "插件"
                ],
                "summary": "更改某一个插件状态",
                "parameters": [
                    {
                        "description": "修改插件状态入参",
                        "name": "object",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/types.PluginStatusParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            }
        },
        "/api/manage/updateResponseStatus": {
            "post": {
                "description": "更改某一个群响应",
                "tags": [
                    "响应"
                ],
                "summary": "更改某一个群响应",
                "parameters": [
                    {
                        "description": "修改群响应入参",
                        "name": "object",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/types.ResponseStatusParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "types.AllPluginParams": {
            "description": "GetAllPlugin的入参",
            "type": "object",
            "properties": {
                "groupId": {
                    "description": "群id, gid\u003e0为群聊,gid\u003c0为私聊,gid=0为全部群聊",
                    "type": "integer"
                }
            }
        },
...
        "types.BotParams": {
            "description": "GetGroupList,GetFriendList的入参",
            "type": "object",
            "properties": {
                "selfId": {
                    "description": "机器人qq",
                    "type": "integer"
                }
            }
        },
        "types.DeleteGroupParams": {
            "description": "退群或删除好友的入参",
            "type": "object",
            "properties": {
                "groupId": {
                    "description": "群id, gid\u003e0为群聊,gid\u003c0为私聊,gid=0为全部群聊",
                    "type": "integer"
                },
                "selfId": {
                    "description": "机器人qq",
                    "type": "integer"
                }
            }
        },
...
        "types.RequestVo": {
            "description": "请求返回",
            "type": "object",
            "properties": {
                "comment": {
                    "description": "注释",
                    "type": "string"
                },
                "flag": {
                    "description": "请求flag",
                    "type": "string"
                },
                "groupId": {
                    "description": "群id",
                    "type": "integer"
                },
                "groupName": {
                    "description": "群名",
                    "type": "string"
                },
                "nickname": {
                    "description": "昵称",
                    "type": "string"
                },
                "requestType": {
                    "description": "请求类型",
                    "type": "string"
                },
                "selfId": {
                    "description": "机器人qq",
                    "type": "integer"
                },
                "subType": {
                    "description": "请求子类型",
                    "type": "string"
                },
                "userId": {
                    "description": "用户id",
                    "type": "integer"
                }
            }
        },
        "types.Response": {
            "description": "包装返回体",
            "type": "object",
            "properties": {
                "code": {
                    "description": "错误码",
                    "type": "integer"
                },
                "message": {
                    "description": "错误信息",
                    "type": "string"
                },
                "result": {
                    "description": "数据"
                },
                "type": {
                    "description": "待定",
                    "type": "string"
                }
            }
        },
...
    }
}